requirements.txt
requests
python-dateutil

./linkedin-api